sw.js: Do not touch the request if it's not GET
I don't know how I didn't catch this earlier but
basically the refresh button never worked at all
untill now. Because the service worker was
intercepting the POST request to the tarkov.dev api
and I don't know _what_ it was doing to it, but it was
failing. And very interestingly there is no way to know
why it is failing. Firefox just says NS_ERROR_INTERCEPTION_FAILED
which just means the service worker failed to get it. But why?
nobody knows, and quite likely nobody will ever know. Marvelous.